+++ software/iob_clint_timer.c
@@ -6,7 +6,7 @@
 
 */
 
-#include "timer.h"
+#include "iob_clint_timer.h"
 
 void mtimer_set_raw_time_cmp(uint64_t clock_offset) {
     // First of all set
